+++ java/src/boolwidth/heuristics/PartialDecompositionHeuristic.java
@@ -5,7 +5,7 @@ public interface PartialDecompositionHeuristic<V, E> {
 	/**
 	 * Search for best partial decomposition
 	 */
-	public int runHeuristic(LSDecomposition<V, E> decomposition,
-			VertexSplit<V> top, int upper_bound);
+	public long runHeuristic(LSDecomposition<V, E> decomposition,
+			VertexSplit<V> top, long upper_bound);
 
 }
